Convert mail_test to pytest
