+++ tests/CMakeLists.txt
@@ -42,7 +42,6 @@ if(RUN_CORE_TESTS)
   add_python_test(install)
   add_python_test(item)
   add_python_test(logging_config)
-  add_python_test(mail)
   add_python_test(model_singleton)
   add_python_test(routetable)
   add_python_test(notification TIMEOUT 50)
